+++ core/viewsets/__init__.py
@@ -1,3 +1,3 @@
-from core.viewsets.register import RegisterViewSet
-from core.viewsets.login import LoginViewSet
-from core.viewsets.refresh import RefreshViewSet
+from core.viewsets.auth.register import RegisterViewSet
+from core.viewsets.auth.login import LoginViewSet
+from core.viewsets.auth.refresh import RefreshViewSet
